pyproject.toml
[project]
name = "docrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.0",
    "groq>=0.36.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=0.6.5",
    "openai>=2.8.1",
    "pydantic>=2.12.5",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "uuid>=1.30",
    "wikipedia>=1.4.0",
]
